fix(codegen/golang): Use int8 for MySQL TINYINT

+++ internal/endtoend/testdata/vet_explain/mysql/db/models.go
@@ -110,7 +110,7 @@ type Debug struct {
 	Cdouble          float64
 	Cdec             string
 	Cfixed           string
-	Ctinyint         int32
+	Ctinyint         int8
 	Cbool            bool
 	Cmediumint       int32
 	Cbit             interface{}
